Implemented http_header Map abstraction
Implemented a bunch of functions to use http_header in a more
conventional manner, as a sort of Map data structure, and replaced
manual usage(i.e. directly accessing it's variables) of the header
at request handling with header function calls.

http_header_t - A Map like data structure of headers
The headers are stored in the struct inside two arrays: fields[], and
values[], that store the names of the headers and their respective
values. The values fieldCount and arraySize represent the Map's current
amount of elements and the arrays' total allocated size respectively.
The keys(fields) are stored as is (keys are case irrelevant) and not
sorted or hashed as one would expect from a map this is becuase there's
never really a large enough amount of headers to demand sorting
them(There's gonna be like what? Max 20? Usually 5-10 headers) and so
this leaves the implementation extremely simple. New name:value pairs
are simply appended to the end of the array.
The map owns and handles it's own memory, every call to the add funcion
duplicates the string pointers and holds it's own copy of the strings
and frees them when it's destroyed, the map never transfers ownership of
it's memory to a user or takes ownership of user's memory.

include/errors.h:
 - Added new header file for error codes and CHECK macros.

include/http.h:
 - Added for http_* struct types a typedef to http_*_t.
 - Added a bunch of header functions for http_header_t:
   - headerNew: Allocate and return a new header struct. Paired
     with headerFree.
   - headerInit: Initialize a header struct's internal values.
     Paired with headerDestroy.
   - headerFree: Free header struct pointer.
   - headerDestroy: Free header struct's internal memory.
   - headerAdd: Add a name:value pair to the header map.
   - headerFind: Find name:value pair index by name.
   - headerGetByName: Get value by name.
   - headerGetByIndex: Get value by index.
   - headerGetNameByIndex: Get name by index.
   - headerLength: Amount of name:value pairs in the header map.

src/http.c:
 - Implemented afforementioned header functions.
 - Implemented an strcmpi function for case irrelvant string
   comparisons for the Find/Get functions.
 - Replaced manual usages of the header's internals by the request
   handlers with calls to header* functions.

src/server.c:
 - Use new header methods.